maxed out timer resolution

+++ graprof/trace.c
@@ -12,9 +12,9 @@ trace_enter (const char *line)
 {
   unsigned long func;
   unsigned long caller;
-  unsigned long time;
+  unsigned long long time;
 
-  int res = sscanf(line, "e 0x%lx 0x%lx %lu", &func, &caller, &time);
+  int res = sscanf(line, "e 0x%lx 0x%lx %llu", &func, &caller, &time);
   assert_inner(res == 3, "sscanf");
 
   timeline_event *e = timeline_push_event(TIMELINE_EVENT_ENTER, time, caller);
@@ -30,9 +30,9 @@ trace_exit (const char *line)
 {
   unsigned long func;
   unsigned long caller;
-  unsigned long time;
+  unsigned long long time;
 
-  unsigned int res = sscanf(line, "x 0x%lx 0x%lx %lu", &func, &caller, &time);
+  unsigned int res = sscanf(line, "x 0x%lx 0x%lx %llu", &func, &caller, &time);
   assert_inner(res == 3, "sscanf");
 
   timeline_event *e = timeline_push_event(TIMELINE_EVENT_EXIT, time, caller);
@@ -49,9 +49,9 @@ trace_malloc (const char *line)
   unsigned int size;
   unsigned long caller;
   unsigned long result;
-  unsigned long time;
+  unsigned long long time;
 
-  unsigned int res = sscanf(line, "+ %u 0x%lx 0x%lx %lu", &size, &caller, &result, &time);
+  unsigned int res = sscanf(line, "+ %u 0x%lx 0x%lx %llu", &size, &caller, &result, &time);
   assert_inner(res == 4, "sscanf");
 
   timeline_event *e = timeline_push_event(TIMELINE_EVENT_MALLOC, time, caller);
@@ -70,9 +70,9 @@ trace_realloc (const char *line)
   unsigned int size;
   unsigned long caller;
   unsigned long result;
-  unsigned long time;
+  unsigned long long time;
 
-  unsigned int res = sscanf(line, "* 0x%lx %u 0x%lx 0x%lx %lu", &ptr, &size, &caller, &result, &time);
+  unsigned int res = sscanf(line, "* 0x%lx %u 0x%lx 0x%lx %llu", &ptr, &size, &caller, &result, &time);
   assert_inner(res == 5, "sscanf");
 
   timeline_event *e = timeline_push_event(TIMELINE_EVENT_REALLOC, time, caller);
@@ -90,9 +90,9 @@ trace_free (const char *line)
 {
   unsigned long ptr;
   unsigned long caller;
-  unsigned long time;
+  unsigned long long time;
 
-  unsigned int res = sscanf(line, "- 0x%lx 0x%lx %lu", &ptr, &caller, &time);
+  unsigned int res = sscanf(line, "- 0x%lx 0x%lx %llu", &ptr, &caller, &time);
   assert_inner(res == 3, "sscanf");
 
   timeline_event *e = timeline_push_event(TIMELINE_EVENT_FREE, time, caller);
